refactor experiment details to use tabs
