+++ app/build.gradle
@@ -51,6 +51,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-installations:16.3.5'
     implementation 'com.google.firebase:firebase-firestore:22.1.1'
     implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.0.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.0.1'
     testImplementation 'junit:junit:4.+'
