+++ app/src/androidTest/java/com/T05/krowdtrialz/ExperimentDetailsOwnerActivityTest.java
@@ -1,21 +1,14 @@
 package com.T05.krowdtrialz;
 
-import android.app.Activity;
-import android.os.Environment;
-import android.util.Log;
-
 import androidx.test.platform.app.InstrumentationRegistry;
 import androidx.test.rule.ActivityTestRule;
 
-import com.T05.krowdtrialz.ui.ExperimentDetailsOwnerActivity;
 import com.robotium.solo.Solo;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-
-import java.io.File;
 
 public class ExperimentDetailsOwnerActivityTest {
     private Solo solo;
