+++ app/src/main/AndroidManifest.xml
@@ -26,8 +26,7 @@
         <activity
             android:name=".ui.contributors.ContributorsActivity"
             android:parentActivityName=".ui.ExperimentDetailsOwnerActivity" />
-        <activity android:name=".ui.ExperimentDetailsNonOwnerActivity" />
-        <activity android:name=".ui.ExperimentDetailsOwnerActivity" />
+        <activity android:name=".ui.experimentDetails.ExperimentDetailsActivity" />
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name" />
