+++ app/src/main/java/com/T05/krowdtrialz/ui/subscribed/SubscribedFragment.java
@@ -1,13 +1,10 @@
 package com.T05.krowdtrialz.ui.subscribed;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
@@ -24,12 +21,8 @@ import androidx.lifecycle.ViewModelProvider;
 import com.T05.krowdtrialz.R;
 import com.T05.krowdtrialz.model.experiment.Experiment;
 import com.T05.krowdtrialz.model.user.User;
-import com.T05.krowdtrialz.ui.ExperimentDetailsNonOwnerActivity;
-import com.T05.krowdtrialz.ui.ExperimentDetailsOwnerActivity;
 import com.T05.krowdtrialz.util.Database;
 import com.T05.krowdtrialz.util.ExperimentList;
-
-import org.apache.commons.math3.analysis.function.Exp;
 
 import java.util.ArrayList;
