+++ app/src/main/res/layout/activity_experiment_details.xml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ui.experimentDetails.ExperimentDetailsActivity">
+
+    <com.google.android.material.textview.MaterialTextView
+        android:id="@+id/owner_textView_experiment"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="12dp"
+        android:layout_marginLeft="12dp"
+        android:layout_marginTop="16dp"
+        android:text="@string/owner_username"
+        app:layout_constraintBottom_toTopOf="@+id/description_textView_experiment"
+        app:layout_constraintEnd_toStartOf="@+id/description_textView_experiment"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.0" />
+
+    <com.google.android.material.textview.MaterialTextView
+        android:id="@+id/description_textView_experiment"
+        android:layout_width="233dp"
+        android:layout_height="106dp"
+        android:layout_marginTop="16dp"
+        android:text="@string/description_non_owner_text"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.9"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <com.google.android.material.checkbox.MaterialCheckBox
+        android:id="@+id/status_checkBox_experiment"
+        android:layout_width="84dp"
+        android:layout_height="44dp"
+        android:layout_marginStart="12dp"
+        android:layout_marginLeft="12dp"
+        android:clickable="false"
+        android:text="@string/status"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/region_textView_experiment" />
+
+    <TextView
+        android:id="@+id/region_textView_experiment"
+        android:layout_width="85dp"
+        android:layout_height="19dp"
+        android:layout_marginStart="12dp"
+        android:layout_marginLeft="12dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/region_non_owner"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/owner_textView_experiment" />
+
+    <Button
+        android:id="@+id/subscribe_button_experiment"
+        android:layout_width="128dp"
+        android:layout_height="50dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/subscribe_non_owner"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.534"
+        app:layout_constraintStart_toEndOf="@+id/add_trials_experiment"
+        app:layout_constraintTop_toBottomOf="@+id/description_textView_experiment" />
+
+    <Button
+        android:id="@+id/add_trials_experiment"
+        android:layout_width="128dp"
+        android:layout_height="50dp"
+        android:layout_marginStart="36dp"
+        android:layout_marginLeft="36dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/add_trials_non_owner"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/description_textView_experiment" />
+
+    <com.google.android.material.tabs.TabLayout
+        android:id="@+id/experiment_tabs"
+        android:layout_width="411dp"
+        android:layout_height="53dp"
+        app:tabTextAppearance="@android:style/TextAppearance.Widget.TabWidget"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/add_trials_experiment"/>
+
+    <FrameLayout
+        android:id="@+id/empty_frame"
+        android:layout_width="412dp"
+        android:layout_height="486dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/experiment_tabs">
+
+    </FrameLayout>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
