+++ app/src/main/res/values/strings.xml
@@ -47,4 +47,6 @@
     <string name="view_q_amp_a_owner"><![CDATA[View Q&A]]></string>
     <string name="subscribed_experiments_title">Subscribed Experiments</string>
     <string name="edit_profile_button_text">Edit Profile</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
